Refactor: Remove Fake AI Chatbot and Consolidate Services
This commit addresses two issues identified in the AI feature review:

1.  **Remove Fake AI Chatbot:** Deleted the non-functional `AIChatbot` and `EnhancedAIChatbot` components and removed all references from the application. This eliminates a misleading UI feature.

2.  **Consolidate AI Services:** Removed the redundant Python `EmbeddingService` to streamline the codebase. The embedding logic was moved directly into the scraper that uses it. This reduces code duplication and improves maintainability.
