+++ job_scraper/scrapers/swissdevjobs_scraper.py
@@ -11,8 +11,52 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../.
 from job_scraper.db.supabase_client import SupabaseClient
 from job_scraper.scrapers.base_scraper import BaseScraper
 from job_scraper.utils.normalize import create_job_hash
+from openai import OpenAI
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+
+class EmbeddingService:
+    """
+    A service to generate text embeddings using OpenAI's API.
+    """
+    def __init__(self):
+        # The OpenAI client will automatically use the OPENAI_API_KEY environment variable.
+        try:
+            self.client = OpenAI()
+            self.model = "text-embedding-ada-002"
+            logging.info("OpenAI client initialized successfully.")
+        except Exception as e:
+            logging.error(f"Failed to initialize OpenAI client: {e}", exc_info=True)
+            self.client = None
+
+    def get_embedding(self, text: str):
+        """
+        Generates an embedding for the given text.
+
+        Args:
+            text: The text to embed.
+
+        Returns:
+            A list of floats representing the embedding, or None if an error occurs.
+        """
+        if not self.client:
+            logging.error("OpenAI client is not available. Cannot generate embedding.")
+            return None
+
+        if not text or not isinstance(text, str):
+            logging.warning("Invalid input text provided for embedding.")
+            return None
+
+        try:
+            # Replace newlines, which can negatively affect performance.
+            text = text.replace("\n", " ")
+            response = self.client.embeddings.create(input=[text], model=self.model)
+            embedding = response.data[0].embedding
+            logging.info(f"Successfully generated embedding for text snippet: '{text[:50]}...'")
+            return embedding
+        except Exception as e:
+            logging.error(f"An error occurred while generating embedding: {e}", exc_info=True)
+            return None
 
 class SwissDevJobsScraper(BaseScraper):
     """
